zz.pl: filtered lines keep their sides
i.e. when filtering by sunstring search lines that were in upper
part of the window stays there, and also lower ones in lower part
of the window. the lines that did not fit above or below the window
appear above or below, respectively, when space appear from filtered
out lines.