Add teebgts: tee(1) from subprocess, with timestamps. subproc exit code
When tee(1) is run in shell pipeline, the exit value of the process
piped to tee(1) is (normally) lost. teebgts runs program itself in
child process, and reports its exit value to the caller...

teebgts logs start and end times, and before each line timestamp
telling the time spent (in milliseconds) running the process.

Introduction to readme to be written...