Add teedots: log to file like teebgts, subprog stdout|err lines as dots
teedots log stdout/stderr of a subprocess to to a file, and just dots
per newline in subprocess stdout/stderr to stdout. Exit value of
subprocess is reported to the caller...

teedots logs start and end times, and before each line timestamp
telling the time spent (in milliseconds) running the process.
Stdout/stderr have (newline and) timestamps after every 64 dots.

Introduction to readme to be written...