tee(bg|do)ts: require '.' as first support to/and support hashbang use
So one can have an executable file starting e.g. with:

    #!./teebgts . logfile /bin/sh -eufx

    echo shell script content

(and same with teedots)

the dot (.) used to recognize when in hashbang line, and to avoid
confusion should one still have older version in use. (if '.' was
not required, older version would execve-loop when used in hashbang
line, as 1st arg was all given in hashbang line and 2nd the binary
itself (that being teebgts or teedots).)