tee(bg|do)ts: requre ofile and command to exist in "hashbang" line
Previously missing command caused teebgts or teedots to be
fork-exec looped until exhaustion.

Attempted to move '.' after filename, then thought it (dot) could
not be given as argument, but it can -- as the file is given as
argument before anything from command line to hashbang-started
executable. As such usage is (IMO) a bit clearer saving these
intermediate copies...