tee(bg|do)ts: moved dot (.) arg from beginning to after ofile
Usage being .../teebgts ofile '.' command [args]

or, #! .../teedots ofile . command [initial-args]

(former on shell command line, latter in hashbang line)

It is IMO clearer to have the "separator" . between ofile
and command -- and as '.' is considered current directory
it cannot be neither file nor executable in file system.

When hashbang line is expanded, system adds the file with
hashbang to argv before any args user may enter so there
is no problem giving '.' as first arg in the command line.