Add (sample) opts--long C code
A piece of source code copyable to a C program -- to do command
line parsing much like getopt_long(3) does it. As this is part
of a program source code (not from a lib), this has consistent
behavior, and is freely editable for feature changes.

The effective part showing features with usage instructions
in the sample is in the last 100 lines of the code.